requirements.txt
streamlit
pandas
matplotlib
openai
openpyxl
tiktoken
python-dotenv